convert jobOutput wrapper function to h2oJobOutput
