+++ src/index.js
@@ -6,7 +6,6 @@ import { flowForm } from './flowForm';
 import { flowPreludeFunction } from './flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
 
-import { jobOutput } from './jobOutput/jobOutput';
 import { routines } from './routines/routines';
 import { localStorage } from './localStorage/localStorage';
 import { knockout } from './knockout/knockout';
@@ -69,5 +68,4 @@ import { flow } from './flow/flow';
     });
   }.call(this));
   routines();
-  jobOutput();
 }).call(this);
